beskar-lib: new io submodule (keyboard)
